Added multithreading to the experiment runner
In order to run experiments faster, I have added the ability to run multiple experiments at once. Preferably, you must be running on a computer with 4+ cores in order to take full advantage of this feature. If you have less cores, you can manually adjust the number of workers. 

This cuts the run time of the experiments down from over 12 minutes to around 4 minutes.